+++ src/main/java/insane96mcp/enhancedai/modules/base/ai/EANearestAttackableTarget.java
@@ -18,7 +18,7 @@ public class EANearestAttackableTarget<T extends LivingEntity> extends ILNearest
     public EANearestAttackableTarget(Mob goalOwnerIn, Class<T> targetClassIn, boolean mustSee, boolean mustReach, TargetingConditions targetingConditions) {
         super(goalOwnerIn, targetClassIn, mustSee, mustReach, null);
         this.targetEntitySelector = targetingConditions;
-        this.targetEntitySelectorXRay = targetingConditions.copy().ignoreLineOfSight().range(this.getFollowXRayDistance());
+        this.targetEntitySelectorXRay = targetingConditions.copy().ignoreLineOfSight();
     }
 
     @Override
@@ -27,14 +27,14 @@ public class EANearestAttackableTarget<T extends LivingEntity> extends ILNearest
             //Try to find the nearest target without xray, then try with xray if the attribute is not 0
             this.nearestTarget = this.mob.level.getNearestEntity(this.targetClass, this.targetEntitySelector, this.mob, this.mob.getX(), this.mob.getEyeY(), this.mob.getZ(), this.getTargetSearchArea(this.getFollowDistance()));
             if (this.nearestTarget == null && this.getFollowXRayDistance() > 0d) {
-                this.nearestTarget = this.mob.level.getNearestEntity(this.targetClass, this.targetEntitySelectorXRay, this.mob, this.mob.getX(), this.mob.getEyeY(), this.mob.getZ(), this.getTargetSearchArea(this.getFollowXRayDistance()));
+                this.nearestTarget = this.mob.level.getNearestEntity(this.targetClass, this.targetEntitySelectorXRay.range(this.getFollowXRayDistance()), this.mob, this.mob.getX(), this.mob.getEyeY(), this.mob.getZ(), this.getTargetSearchArea(this.getFollowXRayDistance()));
             }
         }
         else {
             //Try to find the nearest player without xray, then try with xray if the attribute is not 0
             this.nearestTarget = this.mob.level.getNearestPlayer(this.targetEntitySelector, this.mob, this.mob.getX(), this.mob.getEyeY(), this.mob.getZ());
             if (this.nearestTarget == null && this.getFollowXRayDistance() > 0d) {
-                this.nearestTarget = this.mob.level.getNearestPlayer(this.targetEntitySelectorXRay, this.mob, this.mob.getX(), this.mob.getEyeY(), this.mob.getZ());
+                this.nearestTarget = this.mob.level.getNearestPlayer(this.targetEntitySelectorXRay.range(this.getFollowXRayDistance()), this.mob, this.mob.getX(), this.mob.getEyeY(), this.mob.getZ());
             }
         }
     }